Remove unneeded install from CI
