+++ .travis.yml
@@ -12,9 +12,6 @@ sudo: false
 env:
   - GO111MODULE=on
 
-install:
-  - go install github.com/mattn/go-sqlite3
-
 before_script:
   - mysql -e 'CREATE DATABASE bartlett;'
